added sections

+++ database/migrations/2025_09_26_092451_add_syllabu_id_to_questions_table.php
@@ -12,10 +12,12 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('questions', function (Blueprint $table) {
-            $table->foreignId('syllabu_id')
-                ->after('id') // opcional: define dónde va la columna
-                ->constrained('syllabus')
-                ->onDelete('cascade');
+            if (!Schema::hasColumn('questions', 'syllabu_id')) {
+                $table->foreignId('syllabu_id')
+                    ->after('id') // opcional: define dónde va la columna
+                    ->constrained('syllabus')
+                    ->onDelete('cascade');
+            }
         });
     }
